modificato tipo dato numero

+++ database/migrations/2021_08_31_091032_create_orders_table.php
@@ -20,7 +20,7 @@ class CreateOrdersTable extends Migration
             $table->string('customer_address');
             $table->string('status');
             $table->dateTime('date');
-            $table->smallInteger('total_price');
+            $table->float('total_price');
             $table->timestamps();
         });
     }
